+++ src/notification/views.py
@@ -3,7 +3,7 @@ from rest_framework import status, viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated, IsAdminUser
-from utils.http import PATCH
+from utils.rh_http import PATCH
 
 from notification.models import Notification
 from notification.serializers import NotificationSerializer
