+++ src/reputation/lib.py
@@ -4,7 +4,7 @@ import logging
 import ethereum.utils
 import ethereum.lib
 from researchhub.settings import ASYNC_SERVICE_HOST, WEB3_SHARED_SECRET
-from utils.http import http_request, RequestMethods
+from utils.rh_http import http_request, RequestMethods
 
 WITHDRAWAL_MINIMUM = 5000
 WITHDRAWAL_PER_TWO_WEEKS = 1
