+++ src/utils/permissions.py
@@ -1,6 +1,6 @@
 from rest_framework.exceptions import PermissionDenied  # noqa: F401
 from rest_framework.permissions import BasePermission, SAFE_METHODS
-from utils.http import RequestMethods
+from utils.rh_http import RequestMethods
 
 
 class ReadOnly(BasePermission):
